+++ src/api.js
@@ -136,6 +136,15 @@
     'api',
     'disconnect'
   ]);
+  
+  /*----------------------------------------------------------------
+   * enum for ContactRecording Voice Track config
+   */
+  connect.ContactRecordingVoiceTrackConfig = connect.makeEnum([
+    'ALL',
+    'TO_AGENT',
+    'FROM_AGENT'
+  ]);
 
   /*----------------------------------------------------------------
   * enum ChannelType
@@ -309,7 +318,8 @@
    */
   connect.AgentPermissions = {
     OUTBOUND_CALL: 'outboundCall',
-    VOICE_ID: 'voiceId'
+    VOICE_ID: 'voiceId',
+    CONTACT_RECORDING: 'contactRecording'
   };
 
   /*----------------------------------------------------------------
@@ -515,6 +525,11 @@
 
   Agent.prototype.setConfiguration = function (configuration, callbacks) {
     var client = connect.core.getClient();
+    if (configuration && configuration.agentPreferences && configuration.agentPreferences.LANGUAGE && !configuration.agentPreferences.locale) {
+      // workaround for the inconsistency issue that getAgentConfiguration returns agentPreferences.LANGUAGE but updateAgentConfiguration expects agentPreferences.locale to be set.
+      configuration.agentPreferences.locale = configuration.agentPreferences.LANGUAGE;
+    }
+
     if (configuration && configuration.agentPreferences && !connect.isValidLocale(configuration.agentPreferences.locale)) {
       if (callbacks && callbacks.failure) {
         callbacks.failure(connect.AgentErrorStates.INVALID_LOCALE);
@@ -841,6 +856,10 @@
 
   Contact.prototype.getContactFeatures = function () {
     return this._getData().contactFeatures;
+  };
+
+  Contact.prototype.getChannelContext = function () {
+    return this._getData().channelContext;
   };
 
   Contact.prototype.isSoftphoneCall = function () {
@@ -1180,6 +1199,134 @@
       || connectionType === connect.ConnectionType.MONITORING;
   }
 
+  /*----------------------------------------------------------------
+  * Contact recording
+  */
+  
+  var ContactRecording = function (contactId) {
+    this.contactId = connect.core.getAgentDataProvider().getContactData(contactId) ? contactId : null;
+  }
+
+  ContactRecording.prototype.startContactRecording = function(trackConfig) {
+    var self = this;
+    var client, contactData;
+
+    client = connect.core.getClient();
+    contactData = connect.core.getAgentDataProvider().getContactData(self.contactId);
+    var recordingTrack = (trackConfig && Object.values(connect.ContactRecordingVoiceTrackConfig).includes(trackConfig)) ? trackConfig : connect.ContactRecordingVoiceTrackConfig.ALL;
+    return new Promise(function (resolve, reject) {
+      client.call(connect.AgentAppClientMethods.START_CONTACT_RECORDING, {
+        "initialContactId": contactData.initialContactId || self.contactId,
+        "contactId": self.contactId,
+        "instanceId": connect.core.getAgentDataProvider().getInstanceId(),
+        "voiceRecordingConfiguration": {
+          "voiceRecordingTrack": recordingTrack
+        }
+      }, {
+        success: function (data) {
+          connect.getLog().info("startContactRecording succeeded")
+            .withObject(data).sendInternalLogToServer();
+          resolve(data);
+        },
+        failure: function (err, data) {
+          connect.getLog().error("startContactRecording failed").sendInternalLogToServer()
+            .withObject({
+              err,
+              data
+            });
+          reject(Error("startContactRecording failed", { cause: err }));
+        }
+      })
+    })
+  };
+
+  ContactRecording.prototype.stopContactRecording = function() {
+    var self = this;
+    var client, contactData;
+
+    client = connect.core.getClient();
+    contactData = connect.core.getAgentDataProvider().getContactData(self.contactId);
+    return new Promise(function (resolve, reject) {
+      client.call(connect.AgentAppClientMethods.STOP_CONTACT_RECORDING, {
+        "initialContactId": contactData.initialContactId || self.contactId,
+        "contactId": self.contactId,
+        "instanceId": connect.core.getAgentDataProvider().getInstanceId()
+      }, {
+        success: function (data) {
+          connect.getLog().info("stopContactRecording succeeded")
+            .withObject(data).sendInternalLogToServer();
+          resolve(data);
+        },
+        failure: function (err, data) {
+          connect.getLog().error("stopContactRecording failed").sendInternalLogToServer()
+            .withObject({
+              err,
+              data
+            });
+          reject(Error("stopContactRecording failed", { cause: err }));
+        }
+      })
+    })
+  };
+
+  ContactRecording.prototype.suspendContactRecording = function() {
+    var self = this;
+    var client, contactData;
+
+    client = connect.core.getClient();
+    contactData = connect.core.getAgentDataProvider().getContactData(self.contactId);
+    return new Promise(function (resolve, reject) {
+      client.call(connect.AgentAppClientMethods.SUSPEND_CONTACT_RECORDING, {
+        "initialContactId": contactData.initialContactId || self.contactId,
+        "contactId": self.contactId,
+        "instanceId": connect.core.getAgentDataProvider().getInstanceId()
+      }, {
+        success: function (data) {
+          connect.getLog().info("suspendContactRecording succeeded")
+            .withObject(data).sendInternalLogToServer();
+          resolve(data);
+        },
+        failure: function (err, data) {
+          connect.getLog().error("suspendContactRecording failed").sendInternalLogToServer()
+            .withObject({
+              err,
+              data
+            });
+          reject(Error("suspendContactRecording failed", { cause: err }));
+        }
+      })
+    })
+  };
+
+  ContactRecording.prototype.resumeContactRecording = function() {
+    var self = this;
+    var client, contactData;
+
+    client = connect.core.getClient();
+    contactData = connect.core.getAgentDataProvider().getContactData(self.contactId);
+    return new Promise(function (resolve, reject) {
+      client.call(connect.AgentAppClientMethods.RESUME_CONTACT_RECORDING, {
+        "initialContactId": contactData.initialContactId || self.contactId,
+        "contactId": self.contactId,
+        "instanceId": connect.core.getAgentDataProvider().getInstanceId()
+      }, {
+        success: function (data) {
+          connect.getLog().info("resumeContactRecording succeeded")
+            .withObject(data).sendInternalLogToServer();
+          resolve(data);
+        },
+        failure: function (err, data) {
+          connect.getLog().error("resumeContactRecording failed").sendInternalLogToServer()
+            .withObject({
+              err,
+              data
+            });
+          reject(Error("resumeContactRecording failed"), { cause: err });
+        }
+      })
+    })
+  };
+  
   /*----------------------------------------------------------------
   * Voice authenticator VoiceId
   */
@@ -1583,9 +1730,10 @@
     });
   };
 
-  VoiceId.prototype.checkEnrollmentStatus = function () {
+  VoiceId.prototype.checkEnrollmentStatus = function (callbackOnAudioCollectionComplete) {
     var self = this;
     var pollingTimes = 0;
+    var callbackOnAudioCollectionCompleteHasBeenInvoked = false;
 
     return new Promise(function (resolve, reject) {
       function describe () {
@@ -1596,6 +1744,10 @@
                 resolve(data);
                 break;
               case connect.VoiceIdEnrollmentRequestStatus.IN_PROGRESS:
+                if (!callbackOnAudioCollectionCompleteHasBeenInvoked && typeof callbackOnAudioCollectionComplete === 'function') {
+                  callbackOnAudioCollectionComplete(data);
+                  callbackOnAudioCollectionCompleteHasBeenInvoked = true;
+                }
                 setTimeout(describe, connect.VoiceIdConstants.ENROLLMENT_POLLING_INTERVAL);
                 break;
               case connect.VoiceIdEnrollmentRequestStatus.NOT_ENOUGH_SPEECH:
@@ -1628,7 +1780,7 @@
     });
   };
 
-  VoiceId.prototype.enrollSpeaker = function () {
+  VoiceId.prototype.enrollSpeaker = function (callbackOnAudioCollectionComplete) {
     var self = this;
     self.checkConferenceCall();
     return new Promise(function(resolve, reject) {
@@ -1636,12 +1788,12 @@
         self.getSpeakerStatus().then(function(data) {
           if(data.Speaker && data.Speaker.Status == connect.VoiceIdSpeakerStatus.OPTED_OUT) {
             self.deleteSpeaker().then(function() {
-              self.enrollSpeakerHelper(resolve, reject);
+              self.enrollSpeakerHelper(resolve, reject, callbackOnAudioCollectionComplete);
             }).catch(function(err) {
               reject(err);
             });
           } else {
-            self.enrollSpeakerHelper(resolve, reject);
+            self.enrollSpeakerHelper(resolve, reject, callbackOnAudioCollectionComplete);
           }
         }).catch(function(err) {
           reject(err);
@@ -1652,7 +1804,7 @@
     })
   }
 
-  VoiceId.prototype.enrollSpeakerHelper = function (resolve, reject) {
+  VoiceId.prototype.enrollSpeakerHelper = function (resolve, reject, callbackOnAudioCollectionComplete) {
     var self = this;
     var client = connect.core.getClient();
     var contactData = connect.core.getAgentDataProvider().getContactData(this.contactId);
@@ -1666,7 +1818,7 @@
               connect.getLog().info("enrollSpeaker succeeded").withObject(data).sendInternalLogToServer();
               resolve(data);
             } else {
-              self.checkEnrollmentStatus().then(function(data){
+              self.checkEnrollmentStatus(callbackOnAudioCollectionComplete).then(function(data){
                 connect.getLog().info("enrollSpeaker succeeded").withObject(data).sendInternalLogToServer();
                 resolve(data);
               }).catch(function(err){
@@ -1867,6 +2019,7 @@
    */
   var VoiceConnection = function (contactId, connectionId) {
     this._speakerAuthenticator = new VoiceId(contactId);
+    this._contactRecorder = new ContactRecording(contactId);
     Connection.call(this, contactId, connectionId);
   };
 
@@ -1914,8 +2067,8 @@
     return this._speakerAuthenticator.evaluateSpeaker(startNew);
   }
 
-  VoiceConnection.prototype.enrollSpeakerInVoiceId = function() {
-    return this._speakerAuthenticator.enrollSpeaker();
+  VoiceConnection.prototype.enrollSpeakerInVoiceId = function(callbackOnAudioCollectionComplete) {
+    return this._speakerAuthenticator.enrollSpeaker(callbackOnAudioCollectionComplete);
   }
 
   VoiceConnection.prototype.updateVoiceIdSpeakerId = function(speakerId) {
@@ -1946,6 +2099,39 @@
     }, callbacks);
   };
 
+  VoiceConnection.prototype.startContactRecording = function() {
+    var self = this;
+    self.checkConferenceCall();
+    return this._contactRecorder.startContactRecording();
+  }
+
+  VoiceConnection.prototype.stopContactRecording = function() {
+    var self = this;
+    self.checkConferenceCall();
+    return this._contactRecorder.stopContactRecording();
+  }
+
+  VoiceConnection.prototype.suspendContactRecording = function() {
+    var self = this;
+    self.checkConferenceCall();
+    return this._contactRecorder.suspendContactRecording();
+  }
+
+  VoiceConnection.prototype.resumeContactRecording = function() {
+    var self = this;
+    self.checkConferenceCall();
+    return this._contactRecorder.resumeContactRecording();
+  }
+
+  VoiceConnection.prototype.checkConferenceCall = function(){
+    var self = this;
+    var isConferenceCall = connect.core.getAgentDataProvider().getContactData(self.contactId).connections.filter(function (conn) {
+      return connect.contains(connect.CONNECTION_ACTIVE_STATES, conn.state.type);
+    }).length > 2;
+    if(isConferenceCall){
+      throw new connect.NotImplementedError("VoiceId and Contact Recording are not supported for conference calls");
+    }
+  }
 
   /**
    * @class ChatConnection
@@ -2239,5 +2425,6 @@
   connect.Address = Endpoint;
   connect.SoftphoneError = SoftphoneError;
   connect.VoiceId = VoiceId;
+  connect.ContactRecording = ContactRecording;
 })();