+++ test/unit/connections.spec.js
@@ -1,6 +1,7 @@
 require("../unit/test-setup.js");
 
 describe('Connections API', function () {
+  const sandbox = sinon.createSandbox();
 
   describe('#Chat Connection API', function () {
 
@@ -24,7 +25,7 @@ describe('Connections API', function () {
 
     before(function () {
 
-      connect.core.getAgentDataProvider = sinon.stub().returns({
+      sandbox.stub(connect.core, 'getAgentDataProvider').returns({
         getContactData: () => { return {} },
         _initMediaController: initMediaController,
         getConnectionData: () => {
@@ -40,7 +41,7 @@ describe('Connections API', function () {
 
     after(function () {
       initMediaController.resetHistory();
-      connect.core.getAgentDataProvider.resetBehavior();
+      sandbox.restore();
     });
 
     it('Should create new Chat connection Object given the chat Contact and Connection Id ', function () {
@@ -78,7 +79,7 @@ describe('Connections API', function () {
  
     before(function () {
       sinon.stub(connect.core, 'getClient');
-      connect.core.getAgentDataProvider = sinon.stub().returns({
+      sandbox.stub(connect.core, 'getAgentDataProvider').returns({
         getContactData: () => { return {connections:[{state:{type:"connected"}}]} },
         _initMediaController: initMediaController,
         getConnectionData: () => {
@@ -95,7 +96,7 @@ describe('Connections API', function () {
     after(function () {
       connect.core.getClient.restore();
       initMediaController.resetHistory();
-      connect.core.getAgentDataProvider.resetBehavior();
+      sandbox.restore();
     });
 
     it('Should create new Voice connection Object given the Voice Contact and Connection Id with Speaker Authenticator ', function () {
@@ -105,6 +106,39 @@ describe('Connections API', function () {
       assert.equal(voiceConnection.getMediaType(), connect.MediaType.SOFTPHONE);
       assert.equal(typeof(voiceConnection.getVoiceIdSpeakerId), 'function');
       assert.equal(typeof(voiceConnection.getVoiceIdSpeakerStatus), 'function')
+      assert.equal(typeof(voiceConnection.startContactRecording), 'function');
+      assert.equal(typeof(voiceConnection.stopContactRecording), 'function');
+      assert.equal(typeof(voiceConnection.resumeContactRecording), 'function');
+      assert.equal(typeof(voiceConnection.suspendContactRecording), 'function');
+    });
+
+    describe('startContactRecording', function() {
+      it('Should return StartContactRecording promise with All tracks recording by default on empty or invalid input.', function () {
+        const voiceConnection = new connect.VoiceConnection(contactId, connectionId);
+        var startContactRecording = voiceConnection.startContactRecording('invalid_input');
+        assert.equal(Promise.resolve(startContactRecording), startContactRecording);
+      });
+    });
+    describe('stopContactRecording', function() {
+      it('Should return StopContactRecording promise.', function () {
+        const voiceConnection = new connect.VoiceConnection(contactId, connectionId);
+        var stopContactRecording = voiceConnection.stopContactRecording();
+        assert.equal(Promise.resolve(stopContactRecording), stopContactRecording);
+      });
+    });
+    describe('suspendContactRecording', function() {
+      it('Should return SuspendContactRecording promise.', function () {
+        const voiceConnection = new connect.VoiceConnection(contactId, connectionId);
+        var suspencContactRecording = voiceConnection.suspendContactRecording();
+        assert.equal(Promise.resolve(suspencContactRecording), suspencContactRecording);
+      });
+    });
+    describe('resumeContactRecording', function() {
+      it('Should return ResumeContactRecording promise.', function () {
+        const voiceConnection = new connect.VoiceConnection(contactId, connectionId);
+        var resumeContactRecording = voiceConnection.resumeContactRecording();
+        assert.equal(Promise.resolve(resumeContactRecording), resumeContactRecording);
+      });
     });
 
     describe('getVoiceIdSpeakerId', function() {
@@ -182,7 +216,7 @@ describe('Connections API', function () {
 
     before(function () {
 
-      connect.core.getAgentDataProvider = sinon.stub().returns({
+      sandbox.stub(connect.core, 'getAgentDataProvider').returns({
         getContactData: () => { return {
           initialContactId: taskMediaInfo.initialContactId
         } },
@@ -205,7 +239,7 @@ describe('Connections API', function () {
     });
 
     after(function() {
-      connect.core.getAgentDataProvider.resetBehavior();
+      sandbox.restore();
     });
 
     it('Should create new Task connection Object given the task Contact and Connection Id ', function () {